[IMP] Print any kind of report (custom, qweb-pdf, qweb-html, webkit and etc
