+++ base_report_to_printer/models/__init__.py
@@ -6,5 +6,4 @@ from . import printing_job
 from . import printing_printer
 from . import printing_server
 from . import printing_report_xml_action
-from . import report
 from . import res_users
